requirements.txt
exif==1.6.1
plum-py==0.8.7
python-dotenv==1.0.1
